+++ src/gocept/amqprun/testing.py
@@ -20,25 +20,6 @@ import sys
 import tempfile
 import time
 import unittest
-import zope.component
-import zope.component.globalregistry
-
-
-class SettingsLayer(plone.testing.Layer):
-
-    defaultBases = (plone.testing.zca.LAYER_CLEANUP,)
-
-    def setUp(self):
-        plone.testing.zca.pushGlobalRegistry()
-        settings = dict()
-        zope.component.provideUtility(
-            settings, provides=gocept.amqprun.interfaces.ISettings)
-
-    def tearDown(self):
-        plone.testing.zca.popGlobalRegistry()
-
-
-SETTINGS_LAYER = SettingsLayer()
 
 
 class ZCASandbox(plone.testing.Layer):